backend/requirements.txt
# FastAPI & Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.23
aiosqlite>=0.19.0

# Data Processing
pandas>=2.2.0
openpyxl>=3.1.2
numpy>=2.0.0

# AI/LLM
google-generativeai>=0.8.0
Pillow>=10.1.0

# Template Learning
imagehash>=4.3.1
scikit-learn>=1.5.0
jellyfish>=1.0.0

# Config & Utils
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Packaging
pyinstaller>=6.3.0
pystray>=0.19.5

# Development
pytest>=7.4.0
httpx>=0.25.0
